requirements.txt
anyio==3.6.2
asyncpg==0.27.0
black==23.1.0
cffi==1.15.1
click==8.1.3
cryptography==39.0.2
decouple==0.0.7
dynaconf==3.1.12
ecdsa==0.18.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
jose==1.0.0
jwt==1.3.1
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.1.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
